requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
requests>=2.26.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.2
watchdog>=2.1.6
numpy>=1.21.0
pandas>=1.3.0  # Optional but recommended for data handling
tqdm>=4.62.0  # Optional for progress bars